requirements.txt
requests
requests_oauthlib
python-dotenv
discord.py
tweepy
